Solucionado (creo) el tema del mknod y unlink en Fifo.
El código, aunque medio anidado, debería entenderse.ç

Agrego un poco de documentación también.

+++ fifo/fifo.cpp
@@ -3,28 +3,58 @@
 #include <fcntl.h>
 #include <cerrno>
 #include <cstring>
+#include <cstdio>
 #include <iostream>
-#include <stdio.h>
+
+// Helper functions.
+namespace
+{
+
+bool isFifoFile(const string& pathName)
+{
+	struct stat fileStat;
+	if (stat(pathName.c_str(), &fileStat) == -1)
+		throw FifoError("isFifoFile(): Could not get file info", errno);
+
+	return fileStat.st_mode & S_IFIFO;
+}
+
+} // end namespace.
+
 
 FifoError::FifoError(const string& msj, int errorCode) :
 	runtime_error(msj + " - Error: " + (errorCode ? strerror(errorCode) : ""))
 { }
 
+
 Fifo::Fifo(const string& pathName) : _pathName(pathName)
 {
-	mknod(pathName.c_str(), S_IFIFO|0666, 0);
-	/* if (mknod(pathName.c_str(), S_IFIFO|0666, 0) == -1)
-		throw FifoError("Fifo(): Could not create fifo", errno);
-	*/
+	bool mknodEror = mknod(pathName.c_str(), S_IFIFO | 0666, 0) == -1;
+
+	if (mknodEror)
+	{
+		bool fileAlreadyExists = errno == EEXIST;
+		if (fileAlreadyExists)
+		{
+			if (!isFifoFile(pathName))
+				throw FifoError("Fifo(): File " + pathName + " already exists, "
+			                "but it's not a fifo");
+		}
+		else
+			throw FifoError("Fifo(): Could not create fifo file", errno);
+	}
+
 }
+
 
 Fifo::~Fifo()
 {
-	/*if (unlink(_pathName.c_str()) == -1)
+	bool unlinkError = unlink(_pathName.c_str()) == -1;
+
+	if (unlinkError && errno != ENOENT) // ENOENT = No such file.
 		perror("~Fifo(): Could not unlink the fifo's file");
-	*/
-	unlink(_pathName.c_str());
 }
+
 
 FifoWriter::FifoWriter(const Fifo& fifo)
 {
@@ -34,11 +64,13 @@ FifoWriter::FifoWriter(const Fifo& fifo)
 		                errno);
 }
 
+
 FifoWriter::~FifoWriter()
 {
 	if (close(_fileDescriptor) == -1)
 		perror("~FifoWriter(): Could not close the fifo's file");
 }
+
 
 size_t FifoWriter::write(const void* buffer, size_t size)
 {
@@ -48,6 +80,7 @@ size_t FifoWriter::write(const void* buffer, size_t size)
 
 	return bytes;
 }
+
 
 void FifoWriter::writeFixedSize(const void* buffer, size_t size)
 {
@@ -60,13 +93,14 @@ void FifoWriter::writeFixedSize(const void* buffer, size_t size)
 
 		if (bytes == 0)
 			// Does not use errno.
-			throw FifoError("FifoWriter::writeFixedSize(): could not write to "
+			throw FifoError("FifoWriter::writeFixedSize(): Could not write to "
 			                "fifo");
 
 		totalBytes += bytes;
 		index += bytes;
 	 }
 }
+
 
 FifoReader::FifoReader(const Fifo& fifo)
 {
@@ -76,11 +110,13 @@ FifoReader::FifoReader(const Fifo& fifo)
 		                errno);
 }
 
+
 FifoReader::~FifoReader()
 {
 	if (close(_fileDescriptor) == -1)
 		perror("~FifoReader(): Could not close the fifo's file");
 }
+
 
 size_t FifoReader::read(void* buffer, size_t size)
 {
@@ -90,6 +126,7 @@ size_t FifoReader::read(void* buffer, size_t size)
 
 	return bytes;
 }
+
 
 void FifoReader::readFixedSize(void* buffer, size_t size)
 {
@@ -102,7 +139,7 @@ void FifoReader::readFixedSize(void* buffer, size_t size)
 
 		if (bytes == 0)
 			// Does not use errno.
-			throw FifoError("FifoReader::readFixedSize(): unexpected EOF");
+			throw FifoError("FifoReader::readFixedSize(): Unexpected EOF");
 
 		totalBytes += bytes;
 		index += bytes;
